# Conflicts:
#	src/inverse_modular.rs
#	src/prime_gen.rs

+++ src/utils.rs
@@ -1,12 +1,15 @@
+use num_traits::{One, Unsigned, Zero};
+use std::ops::Rem;
+
 /// Calcule le PGCD de deux nombres
 /// # Arguments
 /// * `a` - Le premier nombre
 /// * `b` - Le deuxième nombre
-pub fn pgcd(a: u64, b: u64) -> u64 {
-    let mut a: u64 = a;
-    let mut b: u64 = b;
+pub fn pgcd(a: u128, b: u128) -> u128 {
+    let mut a: u128 = a;
+    let mut b: u128 = b;
     while b != 0 {
-        let r: u64 = a % b;
+        let r: u128 = a % b;
         a = b;
         b = r;
     }
@@ -19,7 +22,7 @@ pub fn pgcd(a: u64, b: u64) -> u64 {
 /// * `b` - Le deuxième nombre
 /// # Returns
 /// * `true` si les deux nombres sont premiers entre eux, `false` sinon
-pub fn are_coprime(a: u64, b: u64) -> bool {
+pub fn are_coprime(a: u128, b: u128) -> bool {
     return pgcd(a, b) == 1;
 }
 
@@ -81,7 +84,7 @@ pub fn exponential_fast_mod(g: u64, x: u64, modu: u64) -> u64 {
 /// * `q` - Le deuxième nombre
 /// # Returns
 /// * `true` si `e` est premier avec (p − 1)(q − 1), `false` sinon
-pub fn e_is_prime_with(e: u64, p: u64, q: u64) -> bool {
+pub fn e_is_prime_with(e: u128, p: u128, q: u128) -> bool {
     return are_coprime(e, (p - 1) * (q - 1));
 }
